Move out FileManager.java from config package
